Add full validirung client - and server-site

+++ src/actions/serverActions.ts
@@ -3,7 +3,7 @@
 import { prisma } from '@/lib/db';
 import { revalidatePath } from 'next/cache';
 import { Pet } from '@/generated/prisma';
-import { PetEssentials } from '@/types';
+import { petFormSchema, petIdSchema } from '@/lib/schema';
 
 export const getPets = async (): Promise<Pet[]> => {
   try {
@@ -15,13 +15,15 @@ export const getPets = async (): Promise<Pet[]> => {
 };
 
 export const addPet = async (
-  data: PetEssentials,
+  pet: unknown,
 ): Promise<{ message: string } | undefined> => {
+  const validatedPet = petFormSchema.safeParse(pet);
+  if (!validatedPet.success) {
+    return { message: 'Invalid pet data.' };
+  }
+
   try {
-    if (!data.name || !data.ownerName || !data.age) {
-      return { message: 'Some required fields are missing.' };
-    }
-    await prisma.pet.create({ data });
+    await prisma.pet.create({ data: validatedPet.data });
   } catch (error) {
     return {
       message: `Error adding pet: ${error}`,
@@ -32,13 +34,20 @@ export const addPet = async (
 };
 
 export const editPet = async (
-  data: PetEssentials,
-  id: Pet['id'],
+  pet: unknown,
+  id: unknown,
 ): Promise<{ message: string } | undefined> => {
+  const validatedPet = petFormSchema.safeParse(pet);
+  const validatedId = petIdSchema.safeParse(id);
+
+  if (!validatedPet.success || !validatedId.success) {
+    return { message: 'Invalid pet data.' };
+  }
+
   try {
     await prisma.pet.update({
-      where: { id },
-      data,
+      where: { id: validatedId.data },
+      data: validatedPet.data,
     });
   } catch (error) {
     return {
@@ -50,34 +59,21 @@ export const editPet = async (
 };
 
 export const deletePet = async (
-  id: Pet['id'],
+  id: unknown,
 ): Promise<{ message: string } | undefined> => {
+  const validatedId = petIdSchema.safeParse(id);
+
+  if (!validatedId.success) {
+    return { message: 'Invalid pet ID format.' };
+  }
+
   try {
     await prisma.pet.delete({
-      where: { id },
+      where: { id: validatedId.data },
     });
   } catch (error) {
-    console.error('Error deleting pet:', error);
     return { message: `Error deleting pet: ${error}` };
   }
 
   revalidatePath('/app', 'layout');
-};
-
-export const createData = async (
-  formData: FormData,
-): Promise<PetEssentials> => {
-  try {
-    return {
-      name: formData.get('name') as string,
-      ownerName: formData.get('ownerName') as string,
-      imageUrl:
-        (formData.get('imageUrl') as string) ||
-        'https://bytegrad.com/course-assets/react-nextjs/pet-placeholder.png',
-      age: Number(formData.get('age') as string),
-      notes: formData.get('notes') as string,
-    };
-  } catch (error) {
-    throw new Error(`Here are problems.... error: ${error}`);
-  }
 };
